// apps/backend/src/common/interfaces/crud-response.interface.ts
export interface CreatedInterface {
  success: boolean;
  message: string;
}

export interface GetOneDataInterface {
  success: boolean;
  message: string;
  data: Object;
}

export interface GetAllDataInterface {
  success: boolean;
  message: string;
  data: Object;
}

export interface UpdatedInterface {
  success: boolean;
  message: string;
}

export interface DeletedInterface {
  success: boolean;
  message: string;
}